move method to refresh
